Pull in `wes-service` package
