+++ requirements-server.txt
@@ -4,3 +4,4 @@ werkzeug>=2.0,<3
 flask-cors==3.0.10
 gunicorn==20.1.0
 celery>=5.1.0, <6
+wes-service>=4.0.0, <5
